pyproject.toml
[project]
name = "home-watcher"
version = "0.1.0"
description = "Event-driven multi-signal home camera alerting via UniFi Protect"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "Apache-2.0" }
authors = [{ name = "Malin Nordström", email = "mal.nordstrom@gmail.com" }]
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "websockets>=13.0",
    "face-recognition>=1.3",
    "opencv-python-headless>=4.10",
    "pillow>=10.0",
    "numpy>=1.26",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlmodel>=0.0.22",
    "structlog>=24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "ruff>=0.6",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/home_watcher"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "ASYNC"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
